rename underscored to non-underscored
