+++ rules-tests/SwiftMailer/Rector/ClassMethod/SwiftMessageToEmailRector/Fixture/simple_interface.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\Symfony\Tests\SwiftMailer\Rector\ClassMethod\Swift_MessageToEmailRector\Fixture;
+namespace Rector\Symfony\Tests\SwiftMailer\Rector\ClassMethod\SwiftMessageToEmailRector\Fixture;
 
 interface SimpleEmailInterface
 {
@@ -11,7 +11,7 @@ interface SimpleEmailInterface
 -----
 <?php
 
-namespace Rector\Symfony\Tests\SwiftMailer\Rector\ClassMethod\Swift_MessageToEmailRector\Fixture;
+namespace Rector\Symfony\Tests\SwiftMailer\Rector\ClassMethod\SwiftMessageToEmailRector\Fixture;
 
 interface SimpleEmailInterface
 {
